review: move ValueOptions to SDK
